pyproject.toml
[project]
name = "nonebot-plugin-deer-pipe"
version = "0.6.0a2"
description = "A deer-pipe check in plugin for nonebot2"
readme = "README.md"
authors = [{ name = "SNRainiar", email = "rainiar@foxmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.13.3",
    "aiosqlite>=0.20.0",
    "fonttools>=4.61.1",
    "nonebot-plugin-alconna>=0.54.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.2",
    "nonebot-plugin-uninfo>=0.10.1",
    "nonebot2>=2.4.0",
    "pillow>=11.0.0",
    "pytimeparse>=1.1.8",
    "sqlmodel>=0.0.22",
]
license = { text = "AGPL-3.0-or-later" }
classifiers = [
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/SamuNatsu/nonebot-plugin-deer-pipe"
Issues = "https://github.com/SamuNatsu/nonebot-plugin-deer-pipe/issues"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["nonebot-adapter-console>=0.9.0"]
